+++ offload/plugins-nextgen/level_zero/include/L0Device.h
@@ -17,7 +17,7 @@
 
 #include "PerThreadTable.h"
 
-#include "AsyncQueue.h"
+#include "L0Queue.h"
 #include "L0CmdListManager.h"
 #include "L0Context.h"
 #include "L0Program.h"
