+++ offload/plugins-nextgen/level_zero/src/L0Queue.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "AsyncQueue.h"
+#include "L0Queue.h"
 #include "L0Device.h"
 #include "L0Kernel.h"
 #include "L0Plugin.h"
